Renamed the api config file
